backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai>=1.12.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.26.0
httpx>=0.25.0
elevenlabs>=1.0.0

